fixed 9c1c5e0 from https://gitee.com/superFat/ace_ace_engine/pulls/1898 modify code check error

+++ frameworks/core/components/track/rosen_render_arc_track.cpp
@@ -168,7 +168,7 @@ void RosenRenderArcTrack::Paint(RenderContext& context, const Offset& offset)
         double highLightStart = 0.0;
         size_t highLightIndex = 0;
         double ratio = leftToRight_ ? GetTotalRatio() : 1 - GetTotalRatio();
-        for (uint32_t index = colors_.size() - 1; index >= 0; --index) {
+        for (int32_t index = static_cast<int32_t>(colors_.size()) - 1; index >= 0; --index) {
             data.color = colors_[index];
             data.color.ChangeAlpha(UNSELECT_ALPHA);
             currentStart += weights_[index];
